requirements.txt

# ====== NÚCLEO Y MULTIMEDIA ======
numpy==1.22.0
pandas==1.4.4
opencv-python==4.9.0.80
pillow==9.5.0
imutils==0.5.4
pyyaml==6.0

# ====== AUDIO Y PROCESAMIENTO DE VOZ ======
sounddevice
soundfile
pyaudio
webrtcvad==2.0.10
pvporcupine
resampy==0.4.2
pydub==0.25.1
resemblyzer
TTS==0.22.0
openai-whisper

# ====== IA / DEEP LEARNING / MODELOS ======
torch==2.1.2
torchaudio==2.1.2
transformers==4.33.2
cmake==3.25.0
dlib-bin==19.24.2
face-recognition==1.3.0

# ====== BACKEND / API / SERVICIOS WEB ======
fastapi==0.109.2
uvicorn==0.27.1
Flask==2.2.2
flask-cors==3.0.10
jinja2
sqlalchemy==2.0.23
requests
httpx
python-multipart
python-dotenv
paho-mqtt
ollama==0.1.5
pytest==8.0.0

# ====== VALIDACIÓN / CONFIGURACIÓN / UTILIDADES ======
pydantic==2.6.1
aiosqlite==0.19.0

# ====== AUTENTICACIÓN / SEGURIDAD JWT /  OAUTH2  ======
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
